third_party/WebKit/LayoutTests/external/wpt/custom-elements/custom-element-registry/per-global-expected.txt

This is a testharness.js-based test.
PASS Discarding the browsing context must not change window.customElements
FAIL Navigating from the initial about:blank must not replace window.customElementsassert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
FAIL document.open() must replace window.customElementsassert_not_equals: got disallowed value object "[object CustomElementRegistry]"
Harness: the test ran to completion.
